+++ src/pages/Bookmark/_base/BookmarkLayout.jsx
@@ -3,7 +3,7 @@ import React from "react";
 import styled from "styled-components";
 import { CenterLayoutBorder } from "../../../style/Mixin";
 
-import Navbar from "../../../elements/Navbar";
+import Navbar from "../../../components/Navbar";
 // import Post from "../elements/Post";
 
 const BookmarkLayout = () => {
